Rework archethic web routers
