--- lib/archethic_web.ex
@@ -1,92 +1,0 @@
-defmodule ArchethicWeb do
-  @moduledoc false
-
-  def controller do
-    quote do
-      use Phoenix.Controller, namespace: ArchethicWeb
-
-      import Plug.Conn
-      import Phoenix.LiveView.Controller
-      alias ArchethicWeb.Router.Helpers, as: Routes
-    end
-  end
-
-  def explorer_view do
-    quote do
-      use Phoenix.View,
-        root: "lib/archethic_web/explorer/templates",
-        namespace: ArchethicWeb
-
-      # Import convenience functions from controllers
-      import Phoenix.Controller, only: [get_flash: 1, get_flash: 2, view_module: 1]
-
-      # Include shared imports and aliases for views
-      unquote(view_helpers())
-    end
-  end
-
-  def aeweb_view do
-    quote do
-      use Phoenix.View,
-        root: "lib/archethic_web/aeweb/templates",
-        namespace: ArchethicWeb.AEWeb
-
-      # Import convenience functions from controllers
-      import Phoenix.Controller, only: [get_flash: 1, get_flash: 2, view_module: 1]
-
-      # Include shared imports and aliases for views
-      unquote(view_helpers())
-    end
-  end
-
-  def live_view do
-    quote do
-      use Phoenix.LiveView,
-        layout: {ArchethicWeb.LayoutView, :live}
-
-      unquote(view_helpers())
-    end
-  end
-
-  def live_component do
-    quote do
-      use Phoenix.LiveComponent
-
-      unquote(view_helpers())
-    end
-  end
-
-  defp view_helpers do
-    quote do
-      # Use all HTML functionality (forms, tags, etc)
-      use Phoenix.HTML
-
-      # Import LiveView helpers (live_render, live_component, live_patch, etc)
-      import Phoenix.LiveView.Helpers
-
-      # Import basic rendering functionality (render, render_layout, etc)
-      import Phoenix.View
-
-      import ArchethicWeb.ErrorHelpers
-      import ArchethicWeb.LayoutHelpers
-
-      alias ArchethicWeb.ExplorerRouter.Helpers, as: Routes
-    end
-  end
-
-  def router do
-    quote do
-      use Phoenix.Router
-      import Plug.Conn
-      import Phoenix.Controller
-      import Phoenix.LiveView.Router
-    end
-  end
-
-  @doc """
-  When used, dispatch to the appropriate controller/view/etc.
-  """
-  defmacro __using__(which) when is_atom(which) do
-    apply(__MODULE__, which, [])
-  end
-end
