--- lib/archethic_web/aeweb/controllers/root_controller.ex
@@ -1,11 +1,0 @@
-defmodule ArchethicWeb.AEWeb.RootController do
-  @moduledoc false
-
-  alias ArchethicWeb.AEWeb.WebHostingController
-
-  use ArchethicWeb.AEWeb, :controller
-
-  def index(conn, params) do
-    WebHostingController.web_hosting(conn, params)
-  end
-end
