+++ lib/archethic_web/explorer/explorer_router.ex
@@ -3,12 +3,9 @@ defmodule ArchethicWeb.ExplorerRouter do
 
   alias ArchethicWeb.Explorer
 
-  alias ArchethicWeb.Plug.ThrottleByIPHigh
-  alias ArchethicWeb.Plug.ThrottleByIPandPath
   alias ArchethicWeb.Plug.ThrottleByIPLow
 
   use ArchethicWeb.Explorer, :router
-  import Phoenix.LiveDashboard.Router
 
   pipeline :browser do
     plug(:accepts, ["html"])
@@ -17,19 +14,7 @@ defmodule ArchethicWeb.ExplorerRouter do
     plug(:protect_from_forgery)
     plug(:put_secure_browser_headers)
     plug(ThrottleByIPLow)
-    plug(:put_root_layout, {ArchethicWeb.Explorer.LayoutView, :root})
-  end
-
-  pipeline :api do
-    plug(:accepts, ["json"])
-    plug(ThrottleByIPLow)
-    plug(ArchethicWeb.API.GraphQLContext)
-  end
-
-  pipeline :unrestricted_api do
-    plug(:accepts, ["json"])
-    plug(ThrottleByIPHigh)
-    plug(ThrottleByIPandPath)
+    plug(:put_root_layout, {Explorer.LayoutView, :root})
   end
 
   scope "/", Explorer do
@@ -76,44 +61,6 @@ defmodule ArchethicWeb.ExplorerRouter do
     live("/code/proposals", CodeProposalsLive)
     live("/code/proposal/:address", CodeProposalDetailsLive)
     get("/code/download", CodeController, :download)
-  end
-
-  scope "/api" do
-    pipe_through(:unrestricted_api)
-    get("/web_hosting/:address/*url_path", ArchethicWeb.AEWeb.WebHostingController, :web_hosting)
-  end
-
-  scope "/api" do
-    pipe_through(:api)
-
-    get(
-      "/last_transaction/:address/content",
-      ArchethicWeb.API.TransactionController,
-      :last_transaction_content
-    )
-
-    post("/origin_key", ArchethicWeb.API.OriginKeyController, :origin_key)
-    post("/transaction", ArchethicWeb.API.TransactionController, :new)
-    post("/transaction_fee", ArchethicWeb.API.TransactionController, :transaction_fee)
-
-    post(
-      "/transaction/contract/simulator",
-      ArchethicWeb.API.TransactionController,
-      :simulate_contract_execution
-    )
-
-    forward(
-      "/graphiql",
-      Absinthe.Plug.GraphiQL,
-      schema: ArchethicWeb.API.GraphQLSchema,
-      socket: ArchethicWeb.UserSocket
-    )
-
-    forward(
-      "/",
-      Absinthe.Plug,
-      schema: ArchethicWeb.API.GraphQLSchema
-    )
   end
 
   live_session :settings, session: {ArchethicWeb.WebUtils, :keep_remote_ip, []} do
